% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_data.R
\name{fill_data}
\alias{fill_data}
\title{Populate Column Downwards with a Non-missing Value}
\usage{
fill_data(x, var, row_num)
}
\arguments{
\item{x}{An object of type data.frame}

\item{var}{The column in which to populate values}

\item{row_num}{The first non-missing row number in column \code{var} that contains
the value to be populated downwards}
}
\value{
A data.frame
}
\description{
This function is inspired by \code{data.table}'s \code{rleid()}, and allows
the user to specify starting point(location) of the value to be used to
populate the missing rows of a column.
}
\examples{

# Generate sample data
x <-
  data.frame(
    index = 1:10,
    sex = c(rep("", 2), "Male", rep("", 3), "Female", rep("", 3))
  )

x

# Populate missing rows of a column, based on the last available value
fill_data(
  x = x,
  var = "sex",
  row_num = 3
)


}
\seealso{
\code{data.table::rleid()}
}
\author{
Cornelius Tanui: \email{kiplimocornelius@gmail.com}
}
