% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tables.R
\name{gen_tables}
\alias{gen_tables}
\title{Generate Formatted Tables}
\usage{
gen_tables(data, group_var, var1, var2, var3)
}
\arguments{
\item{data}{An object of type data.frame}

\item{group_var}{the grouping variable, must be non-missing}

\item{var1}{First stratification variable, can be factor or numeric}

\item{var2}{Second stratification variable, can be factor or numeric}

\item{var3}{Third stratification variable, can be factor or numeric
the value to be populated downwards}
}
\value{
A data.frame
}
\description{
This function is a wrapper of \code{table1()}, and is tailored for
three stratification variables
}
\examples{

# Sample data
x <-
  datasets::Titanic |>
  as.data.frame()

# Generate table
gen_tables(
  data = x,
  group_var = x$Sex,
  var1 = x$Class,
  var2 = x$Survived,
  var3 = x$Freq
)
}
\seealso{
\code{table1()}
}
\author{
Cornelius Tanui: \email{kiplimocornelius@gmail.com}
}
