% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindNA.R
\name{naCbind}
\alias{naCbind}
\title{Combine R Objects with different length by Rows or Columns}
\usage{
naCbind(..., deparse.level = 1)
}
\arguments{
\item{...}{A sequence of vector, matrix or data-frame to bind}

\item{deparse.level}{Integer controlling the construction of labels in the case of
non-matrix-like arguments (for the default method):
deparse.level = 0 constructs no labels; the default, deparse.level = 1 or 2 constructs
labels from the argument names, see the 'Value' section below.}
}
\value{
For the default method, a matrix combining the ... arguments column-wise or
 row-wise. (Exception: if there are no inputs or all the inputs are NULL, the value
 is NULL.)

 The type of a matrix result determined from the highest type of any of the inputs
 in the hierarchy raw < logical < integer < double < complex < character < list .
}
\description{
Take a sequence of vector, matrix or data-frame arguments and combine by
columns or rows, respectively. If the length of the R objects differ
they are filled with NAs.
}
\details{
The code is based on the  \code{\link{cbind}()} function.
}
\examples{
# Vector
naCbind(1:10, 1:20)

# Matrix
naCbind(matrix(1:10, ncol = 2), matrix(1:20, ncol = 2))

# Data.frame
naCbind(data.frame(x=1:10, y=1:10), data.frame(c=1:20, d=1:20))

}
\seealso{
\code{\link[base]{cbind}} or \code{\link[base]{rbind}} documentation.
}
