% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsOver.R
\name{pointsOver}
\alias{pointsOver}
\title{To find the if points are within a polygon shapefile given a set of coordinates}
\usage{
pointsOver(
  lon,
  lat,
  lands,
  projection = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{lon}{A numeric vector with longitude information.}

\item{lat}{A numeric vector with latitude information.}

\item{lands}{Layer from which the geometries or attributes are queried. This arguments allows 'SpatialPolygons' or 'SpatialPolygonsDataFrame' class objects.}

\item{projection}{Character string describing a projection and datum in the PROJ.4 format.}
}
\value{
This function return 'TRUE' or 'FALSE' label based on whether the point is on land or not.
}
\description{
To find the if points are within a polygon shapefile given a set of coordinates
}
\examples{
\dontrun{# Create an artificial data set
x = data.table(lat=1:10, lon=1:10)
world = rnaturalearthdata::countries50
x$label = pointsOnLand(x$lon, x$lon, world)
head(x)
}
}
